Release 20240725.1

+++ Net-React.Server/Controllers/AddressController.cs
@@ -5,40 +5,69 @@ using Net_React.Server.Services.Services;
 
 namespace Net_React.Server.Controllers
 {
-    public class AddressController
+    [ApiController]
+    [Route("api/[Controller]")]
+    public class AddressController : ControllerBase
     {
         private readonly IAddressService _addressService;
         public AddressController(IAddressService addressService)
         {
             _addressService = addressService;
         }
-        public async Task<IEnumerable<Address>> GetAll()
+        [HttpGet]
+        public async Task<IList<Address>> GetAll()
         {
             return _addressService.GetAllAddresses();
         }
 
-        [HttpGet]
+        [HttpGet("{id}")]
         public async Task<Address> GetById(int id)
         {
             return _addressService.GetByAddressId(id);
         }
 
-        [HttpGet("userid")]
+        [HttpGet("{userid}")]
         public async Task<Address> GetByUserId(int userId)
         {
             return _addressService.GetByUserId(userId);
         }
-        public async void Add(Address address)
+        [HttpPost]
+
+        public async Task<ActionResult<Address>> Add(Address address)
         {
             _addressService.AddAddress(address);
+            return CreatedAtAction("GetAddress", new { id = address.Id }, address);
         }
-        public async void Update(Address address)
+
+        [HttpPut("{id}")]
+        public async Task<IActionResult> Update(int id, Address address)
         {
-            _addressService.AddAddress(address);
+            if(id != address.Id) return BadRequest();
+            var model = _addressService.GetByAddressId(id);
+            if (model == null)
+            {
+                return NotFound();
+            }
+            else
+            {
+                _addressService.UpdateAddress(address);
+                return Ok(address);
+            }
         }
-        public async void DeleteById(int id)
+
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> DeleteById(int id)
         {
-            _addressService.DeleteByAddressId(id);
+            var model = _addressService.GetByAddressId(id);
+            if (model == null)
+            {
+                return NotFound();
+            }
+            else
+            {
+                _addressService.DeleteByAddressId(id);
+                return Ok("Deleted " + id);
+            }
         }
     }
 }
